implement new entrypoint for data pipeline, simplify python files
- implement audiochunk happy path entrypoint
- implement happy path for transcripts
- clean up preprocess_data transcript function
- clean up ibm transcript service
