+++ mentors/requirements.txt
@@ -1,13 +1,13 @@
 certifi==2019.6.16
 chardet==3.0.4
 ffmpy==0.2.2
+gdown==3.8.1
 idna==2.8
 numpy==1.16.4
 pandas==0.24.2
 python-dateutil==2.8.0
 pytz==2019.1
 requests==2.22.0
-six==1.12.0
 urllib3==1.25.3
 watson-developer-cloud==2.10.1
 websocket-client==0.48.0
